functions/message-processor/requirements.txt
functions-framework==3.4.0
google-cloud-pubsub==2.18.3
google-cloud-firestore==2.12.0